requirements.txt
# GitUp Requirements
# Core dependencies
click>=8.0.0
GitPython>=3.1.30
PyYAML>=6.0
rich>=12.0.0
pathspec>=0.9.0

# GitGuard integration
gitguard>=1.0.2

# Template processing
jinja2>=3.1.0

# Configuration management
python-dotenv>=0.19.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0